moved files into our own TopDown directory
